refactor FAQSingleEntry is separated from FAQEntryList in its own file
