+++ frontend/src/pages/faq/FAQEntryList.tsx
@@ -1,19 +1,12 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
-import {
-  Box,
-  Button,
-  Grid,
-  IconButton,
-  Paper,
-  Typography,
-} from '@mui/material';
-import { Link, PeopleAlt } from '@mui/icons-material';
+import { Box, Button, Typography } from '@mui/material';
+import { PeopleAlt } from '@mui/icons-material';
 
-import { useSnackbar } from 'notistack';
 import { FAQEntry } from 'src/services/openapi';
-import { SNACKBAR_DYNAMIC_FEEDBACK_MS } from 'src/utils/notifications';
+
+import FAQSingleEntry from 'src/pages/faq/FAQSingleEntry';
 
 /**
  * The Entry List section of the FAQ page.
@@ -26,73 +19,11 @@ import { SNACKBAR_DYNAMIC_FEEDBACK_MS } from 'src/utils/notifications';
  */
 const FAQEntryList = ({ entries }: { entries: Array<FAQEntry> }) => {
   const { t } = useTranslation();
-  const { enqueueSnackbar } = useSnackbar();
-
-  const copyUriToClipboard = (
-    event: React.MouseEvent<HTMLElement>,
-    anchor: string
-  ) => {
-    const uri = `${location.protocol}//${location.host}${location.pathname}#${anchor}`;
-    navigator.clipboard.writeText(uri);
-    enqueueSnackbar(t('faqEntryList.questionURLCopied'), {
-      variant: 'success',
-      autoHideDuration: SNACKBAR_DYNAMIC_FEEDBACK_MS,
-    });
-  };
 
   return (
     <>
       {entries.map((entry) => {
-        const answerParagraphs = entry.answer.split('\n');
-        return (
-          <Box key={`q_${entry.name}`} mb={4}>
-            <Paper square>
-              <Box p={2} pb="1px">
-                <Grid container>
-                  <Grid item xs={11}>
-                    <Typography
-                      id={entry.name}
-                      variant="h4"
-                      gutterBottom
-                      // Match the IconButton padding to align the two items
-                      // nicely, even when the title fits on several lines.
-                      sx={{ pt: 1 }}
-                    >
-                      {entry.question}
-                    </Typography>
-                  </Grid>
-                  <Grid item xs={1}>
-                    <Box
-                      display="flex"
-                      alignItems="center"
-                      justifyContent="flex-end"
-                    >
-                      <IconButton
-                        aria-label="Copy URI to clipboard"
-                        onClick={(event) => {
-                          copyUriToClipboard(event, entry.name);
-                        }}
-                      >
-                        <Link />
-                      </IconButton>
-                    </Box>
-                  </Grid>
-                </Grid>
-
-                {/* An answer can be composed of several paragraphs. */}
-                {answerParagraphs.map((paragraph, index) => (
-                  <Typography
-                    key={`$a_{entry.name}_p${index}`}
-                    paragraph
-                    textAlign="justify"
-                  >
-                    {paragraph}
-                  </Typography>
-                ))}
-              </Box>
-            </Paper>
-          </Box>
-        );
+        return <FAQSingleEntry key={`q_${entry.name}`} entry={entry} />;
       })}
 
       {/* Always display an extra entry to explain how to ask more questions. */}
